Introduced transmit method which returns retry, done or abort status of the process. Rspec unit tests are conceptually reorganized, improved.

+++ lib/logstash/utils/load_balancer.rb
@@ -72,7 +72,7 @@ class LoadBalancer
     end
 
     def mark_error
-      @last_error = Time.now.to_f
+      @last_error = Time.now.to_i
     end
   end
 end
